Added reference to class in each block
Now each block has a constructor which take the class itself as an
argument, this allow to store all of the static content of the block
within the class and still have access to it in the Block class.

Also implemented a way to specified the class to use in the json
definition of  the block even if the javascript code is minified.